+++ .gitignore
@@ -1,5 +1,6 @@
 *.pyc
 .DS_Store
+*.sdf
 
 profile.json
 .vscode
