+++ gypsum_dl/steps/conf/prepare.py
@@ -2,11 +2,9 @@
 Runs the 3D preparation process.
 """
 
-from gypsum_dl.steps.conf.Convert2DTo3D import convert_2d_to_3d
-from gypsum_dl.steps.conf.GenerateAlternate3DNonaromaticRingConfs import (
-    generate_alternate_3d_nonaromatic_ring_confs,
-)
-from gypsum_dl.steps.conf.Minimize3D import minimize_3d
+from gypsum_dl.steps.conf.convert import convert_2d_to_3d
+from gypsum_dl.steps.conf.minimize import minimize_3d
+from gypsum_dl.steps.conf.rings import generate_alternate_3d_nonaromatic_ring_confs
 
 
 def prepare_3d(contnrs, params):
