+++ gypsum_dl/steps/io/output.py
@@ -3,9 +3,9 @@ The proccess_output definition determines which formats are saved to the
 disk (output).
 """
 
-from gypsum_dl.steps.io.SaveToPDB import convert_sdfs_to_PDBs
-from gypsum_dl.steps.io.SaveToSDF import save_to_sdf
-from gypsum_dl.steps.io.Web2DOutput import web_2d_output
+from gypsum_dl.steps.io.to_html import web_2d_output
+from gypsum_dl.steps.io.to_pdb import convert_sdfs_to_PDBs
+from gypsum_dl.steps.io.to_sdf import save_to_sdf
 
 
 def proccess_output(contnrs, params):
